Add files via upload
This is a final year project of my Bsc Honours in Statistics , where I conducted an in-depth analysis of 2011 CENSUS DATA OF INDIA & PREDICTION OF 2021 & 2031 CENSUS. This project allowed me to apply a range of statistical methods and tools to derive actionable insights and predictions.

🔍 Project Overview:

Objective: TO ANALYSIS OF 2011 CENSUS DATA OF INDIA & PREDICTION OF 2021 & 2031 CENSUS

Methods Used:

ANOVA & Linear Regression : Linear regression is a powerful tool for predicting census data outcomes and the ANOVA table helps in understanding the significance of each predictor in the model. The combination of these techniques allows for a prediction of census-related variables.

Statistical Graphs : To visualize trends and relationships within the data.

Tools & Technologies:

R : For advanced statistical analysis and data manipulation.